separate parser and target type, add parameterized query parser and type
